+++ neo/framework/async/AsyncServer.cpp
@@ -769,7 +769,7 @@ void idAsyncServer::DropClient( int clientNum, const char *reason ) {
 	if ( client.clientState >= SCS_PUREWAIT && clientNum != localClientNum ) {
 		msg.Init( msgBuf, sizeof( msgBuf ) );
 		msg.WriteByte( SERVER_RELIABLE_MESSAGE_DISCONNECT );
-		msg.WriteLong( clientNum );
+		msg.WriteInt( clientNum );
 		msg.WriteString( reason );
 		for ( i = 0; i < MAX_ASYNC_CLIENTS; i++ ) {
 			// clientNum so SCS_PUREWAIT client gets it's own disconnect msg
@@ -918,7 +918,7 @@ void idAsyncServer::SendUserInfoBroadcast( int userInfoNum, const idDict &info, 
 	}
 
 #if ID_CLIENTINFO_TAGS
-	msg.WriteLong( sessLocal.mapSpawnData.userInfo[userInfoNum].Checksum() );
+	msg.WriteInt( sessLocal.mapSpawnData.userInfo[userInfoNum].Checksum() );
 	common->DPrintf( "broadcast for client %d: 0x%x\n", userInfoNum, sessLocal.mapSpawnData.userInfo[userInfoNum].Checksum() );
 	sessLocal.mapSpawnData.userInfo[userInfoNum].Print();
 #endif
@@ -976,7 +976,7 @@ void idAsyncServer::SendUserInfoToClient( int clientNum, int userInfoNum, const 
 	msg.WriteBits( 0, 1 );
 
 #if ID_CLIENTINFO_TAGS
-	msg.WriteLong( 0 );
+	msg.WriteInt( 0 );
 	common->DPrintf( "user info %d to client %d: NULL base\n", userInfoNum, clientNum );
 #endif
 
@@ -1039,7 +1039,7 @@ void idAsyncServer::SendApplySnapshotToClient( int clientNum, int sequence ) {
 
 	msg.Init( msgBuf, sizeof( msgBuf ) );
 	msg.WriteByte( SERVER_RELIABLE_MESSAGE_APPLYSNAPSHOT );
-	msg.WriteLong( sequence );
+	msg.WriteInt( sequence );
 
 	SendReliableMessage( clientNum, msg );
 }
@@ -1068,7 +1068,7 @@ bool idAsyncServer::SendEmptyToClient( int clientNum, bool force ) {
 	}
 
 	msg.Init( msgBuf, sizeof( msgBuf ) );
-	msg.WriteLong( gameInitId );
+	msg.WriteInt( gameInitId );
 	msg.WriteByte( SERVER_UNRELIABLE_MESSAGE_EMPTY );
 
 	client.channel.SendMessage( serverPort, serverTime, msg );
@@ -1102,9 +1102,9 @@ bool idAsyncServer::SendPingToClient( int clientNum ) {
 	}
 
 	msg.Init( msgBuf, sizeof( msgBuf ) );
-	msg.WriteLong( gameInitId );
+	msg.WriteInt( gameInitId );
 	msg.WriteByte( SERVER_UNRELIABLE_MESSAGE_PING );
-	msg.WriteLong( realTime );
+	msg.WriteInt( realTime );
 
 	client.channel.SendMessage( serverPort, serverTime, msg );
 
@@ -1134,10 +1134,10 @@ void idAsyncServer::SendGameInitToClient( int clientNum ) {
 	}
 
 	msg.Init( msgBuf, sizeof( msgBuf ) );
-	msg.WriteLong( gameInitId );
+	msg.WriteInt( gameInitId );
 	msg.WriteByte( SERVER_UNRELIABLE_MESSAGE_GAMEINIT );
-	msg.WriteLong( gameFrame );
-	msg.WriteLong( gameTime );
+	msg.WriteInt( gameFrame );
+	msg.WriteInt( gameTime );
 	msg.WriteDeltaDict( sessLocal.mapSpawnData.serverInfo, NULL );
 	client.gameInitSequence = client.channel.SendMessage( serverPort, serverTime, msg );
 }
@@ -1169,11 +1169,11 @@ bool idAsyncServer::SendSnapshotToClient( int clientNum ) {
 
 	// write the snapshot
 	msg.Init( msgBuf, sizeof( msgBuf ) );
-	msg.WriteLong( gameInitId );
+	msg.WriteInt( gameInitId );
 	msg.WriteByte( SERVER_UNRELIABLE_MESSAGE_SNAPSHOT );
-	msg.WriteLong( client.snapshotSequence );
-	msg.WriteLong( gameFrame );
-	msg.WriteLong( gameTime );
+	msg.WriteInt( client.snapshotSequence );
+	msg.WriteInt( gameFrame );
+	msg.WriteInt( gameTime );
 	msg.WriteByte( idMath::ClampChar( client.numDuplicatedUsercmds ) );
 	msg.WriteShort( idMath::ClampShort( client.clientAheadTime ) );
 
@@ -1523,7 +1523,7 @@ void idAsyncServer::ProcessChallengeMessage( const netadr_t from, const idBitMsg
 	outMsg.Init( msgBuf, sizeof( msgBuf ) );
 	outMsg.WriteShort( CONNECTIONLESS_MESSAGE_ID );
 	outMsg.WriteString( "challengeResponse" );
-	outMsg.WriteLong( challenges[i].challenge );
+	outMsg.WriteInt( challenges[i].challenge );
 	outMsg.WriteShort( serverId );
 	outMsg.WriteString( cvarSystem->GetCVarString( "fs_game_base" ) );
 	outMsg.WriteString( cvarSystem->GetCVarString( "fs_game" ) );
@@ -1542,9 +1542,9 @@ void idAsyncServer::ProcessChallengeMessage( const netadr_t from, const idBitMsg
 			outMsg.BeginWriting();
 			outMsg.WriteShort( CONNECTIONLESS_MESSAGE_ID );
 			outMsg.WriteString( "srvAuth" );
-			outMsg.WriteLong( ASYNC_PROTOCOL_VERSION );
+			outMsg.WriteInt( ASYNC_PROTOCOL_VERSION );
 			outMsg.WriteNetadr( from );
-			outMsg.WriteLong( -1 ); // this identifies "challenge" auth vs "connect" auth
+			outMsg.WriteInt( -1 ); // this identifies "challenge" auth vs "connect" auth
 			// protocol 1.37 addition
 			outMsg.WriteByte( fileSystem->RunningD3XP() );
 			serverPort.SendPacket( idAsyncNetwork::GetMasterAddress(), outMsg.GetData(), outMsg.GetSize() );
@@ -1579,12 +1579,12 @@ bool idAsyncServer::SendPureServerMessage( const netadr_t to, int OS ) {
 
 	i = 0;
 	while ( serverChecksums[ i ] ) {
-		outMsg.WriteLong( serverChecksums[ i++ ] );
+		outMsg.WriteInt( serverChecksums[ i++ ] );
 	}
-	outMsg.WriteLong( 0 );
+	outMsg.WriteInt( 0 );
 
 	// write the pak checksum for game code
-	outMsg.WriteLong( gamePakChecksum );
+	outMsg.WriteInt( gamePakChecksum );
 
 	serverPort.SendPacket( to, outMsg.GetData(), outMsg.GetSize() );
 	return true;
@@ -1614,14 +1614,14 @@ bool idAsyncServer::SendReliablePureToClient( int clientNum ) {
 	msg.Init( msgBuf, sizeof( msgBuf ) );
 	msg.WriteByte( SERVER_RELIABLE_MESSAGE_PURE );
 
-	msg.WriteLong( gameInitId );
+	msg.WriteInt( gameInitId );
 
 	i = 0;
 	while ( serverChecksums[ i ] ) {
-		msg.WriteLong( serverChecksums[ i++ ] );
+		msg.WriteInt( serverChecksums[ i++ ] );
 	}
-	msg.WriteLong( 0 );
-	msg.WriteLong( gamePakChecksum );
+	msg.WriteInt( 0 );
+	msg.WriteInt( gamePakChecksum );
 
 	SendReliableMessage( clientNum, msg );
 
@@ -1751,9 +1751,9 @@ void idAsyncServer::ProcessConnectMessage( const netadr_t from, const idBitMsg &
 				outMsg.Init( msgBuf, sizeof( msgBuf ) );
 				outMsg.WriteShort( CONNECTIONLESS_MESSAGE_ID );
 				outMsg.WriteString( "srvAuth" );
-				outMsg.WriteLong( ASYNC_PROTOCOL_VERSION );
+				outMsg.WriteInt( ASYNC_PROTOCOL_VERSION );
 				outMsg.WriteNetadr( from );
-				outMsg.WriteLong( clientId );
+				outMsg.WriteInt( clientId );
 				outMsg.WriteString( guid );
 				// protocol 1.37 addition
 				outMsg.WriteByte( fileSystem->RunningD3XP() );
@@ -1785,8 +1785,8 @@ void idAsyncServer::ProcessConnectMessage( const netadr_t from, const idBitMsg &
 		outMsg.Init( msgBuf, sizeof( msgBuf ) );
 		outMsg.WriteShort( CONNECTIONLESS_MESSAGE_ID );
 		outMsg.WriteString( "print" );
-		outMsg.WriteLong( SERVER_PRINT_GAMEDENY );
-		outMsg.WriteLong( reply );
+		outMsg.WriteInt( SERVER_PRINT_GAMEDENY );
+		outMsg.WriteInt( reply );
 		outMsg.WriteString( reason );
 		serverPort.SendPacket( from, outMsg.GetData(), outMsg.GetSize() );
 
@@ -1860,10 +1860,10 @@ void idAsyncServer::ProcessConnectMessage( const netadr_t from, const idBitMsg &
 	outMsg.Init( msgBuf, sizeof( msgBuf ) );
 	outMsg.WriteShort( CONNECTIONLESS_MESSAGE_ID );
 	outMsg.WriteString( "connectResponse" );
-	outMsg.WriteLong( clientNum );
-	outMsg.WriteLong( gameInitId );
-	outMsg.WriteLong( gameFrame );
-	outMsg.WriteLong( gameTime );
+	outMsg.WriteInt( clientNum );
+	outMsg.WriteInt( gameInitId );
+	outMsg.WriteInt( gameFrame );
+	outMsg.WriteInt( gameTime );
 	outMsg.WriteDeltaDict( sessLocal.mapSpawnData.serverInfo, NULL );
 
 	serverPort.SendPacket( from, outMsg.GetData(), outMsg.GetSize() );
@@ -2075,8 +2075,8 @@ void idAsyncServer::ProcessGetInfoMessage( const netadr_t from, const idBitMsg &
 	outMsg.Init( msgBuf, sizeof( msgBuf ) );
 	outMsg.WriteShort( CONNECTIONLESS_MESSAGE_ID );
 	outMsg.WriteString( "infoResponse" );
-	outMsg.WriteLong( challenge );
-	outMsg.WriteLong( ASYNC_PROTOCOL_VERSION );
+	outMsg.WriteInt( challenge );
+	outMsg.WriteInt( ASYNC_PROTOCOL_VERSION );
 	outMsg.WriteDeltaDict( sessLocal.mapSpawnData.serverInfo, NULL );
 
 	for ( i = 0; i < MAX_ASYNC_CLIENTS; i++ ) {
@@ -2088,11 +2088,11 @@ void idAsyncServer::ProcessGetInfoMessage( const netadr_t from, const idBitMsg &
 
 		outMsg.WriteByte( i );
 		outMsg.WriteShort( client.clientPing );
-		outMsg.WriteLong( client.channel.GetMaxOutgoingRate() );
+		outMsg.WriteInt( client.channel.GetMaxOutgoingRate() );
 		outMsg.WriteString( sessLocal.mapSpawnData.userInfo[i].GetString( "ui_name", "Player" ) );
 	}
 	outMsg.WriteByte( MAX_ASYNC_CLIENTS );
-	outMsg.WriteLong( fileSystem->GetOSMask() );
+	outMsg.WriteInt( fileSystem->GetOSMask() );
 
 	serverPort.SendPacket( from, outMsg.GetData(), outMsg.GetSize() );
 }
@@ -2578,7 +2578,7 @@ void idAsyncServer::PrintOOB( const netadr_t to, int opcode, const char *string 
 	outMsg.Init( msgBuf, sizeof( msgBuf ) );
 	outMsg.WriteShort( CONNECTIONLESS_MESSAGE_ID );
 	outMsg.WriteString( "print" );
-	outMsg.WriteLong( opcode );
+	outMsg.WriteInt( opcode );
 	outMsg.WriteString( string );
 	serverPort.SendPacket( to, outMsg.GetData(), outMsg.GetSize() );
 }
@@ -2741,7 +2741,7 @@ void idAsyncServer::ProcessDownloadRequestMessage( const netadr_t from, const id
 	outMsg.Init( msgBuf, sizeof( msgBuf ) );
 	outMsg.WriteShort( CONNECTIONLESS_MESSAGE_ID );
 	outMsg.WriteString( "downloadInfo" );
-	outMsg.WriteLong( dlRequest );
+	outMsg.WriteInt( dlRequest );
 	if ( !game->DownloadRequest( Sys_NetAdrToString( from ), challenges[ iclient ].guid, paklist.c_str(), pakbuf ) ) {
 		common->DPrintf( "game: no downloads\n" );
 		outMsg.WriteByte( SERVER_DL_NONE );
@@ -2803,7 +2803,7 @@ void idAsyncServer::ProcessDownloadRequestMessage( const netadr_t from, const id
 				tmpMsg.WriteByte( SERVER_PAK_YES );
 				tmpMsg.WriteString( pakNames[ i ] );
 				tmpMsg.WriteString( pakURLs[ i ] );
-				tmpMsg.WriteLong( dlSize[ i ] );
+				tmpMsg.WriteInt( dlSize[ i ] );
 			}
 
 			// keep last 5 bytes for an 'end of message' - SERVER_PAK_END and the totalDlSize long
